+++ lib/xnix/compatibility/dbus_runtime_client.rb
@@ -73,6 +73,10 @@ module Xnix
         parse_dictionary(call("GetKDEIntegrationStatus"))
       end
 
+      def kde_application_surface_plan(application_id)
+        parse_dictionary(call("GetKDEApplicationSurfacePlan", application_id))
+      end
+
       def kwin_window_rule_plan(application_id)
         parse_dictionary(call("GetKWinWindowRulePlan", application_id))
       end